% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Centering Datasets}
\usage{
center(data, desired)
}
\arguments{
\item{data}{The numeric vector to be centered}

\item{desired}{The numeric midpoint value around which the data will be centered.}
}
\value{
A new vector containing the original data centered around the desired values
}
\description{
Some more description.
}
\details{
And even more.
}
\examples{
  center(c(1,2,3), 0) # should return [1] -1  0  1
  center(c(1,2,3), 1) # should return [1] 0 1 2
}
