% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescaling Datasets}
\usage{
rescale(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
A new numeric vector of the same length, but rescaled from the input vector to the range of 0 to 1
}
\description{
Rescaling Datasets
}
\examples{
  rescale(c(1,2,3)) # should return [1] 0.0 0.5 1.0
  rescale(c(1,2,3,4,5)) # should return [1] 0.00 0.25 0.50 0.75 1.00
}
